requirements.txt
imageio==2.19.5
matplotlib==3.5.1
numpy==1.23.1
Pillow==9.2.0
